+++ Robust.Shared/Network/NetManager.cs
@@ -12,6 +12,7 @@ using Prometheus;
 using Robust.Shared.Configuration;
 using Robust.Shared.IoC;
 using Robust.Shared.Log;
+using Robust.Shared.Network.Transfer;
 using Robust.Shared.Player;
 using Robust.Shared.Profiling;
 using Robust.Shared.Serialization;
@@ -112,6 +113,7 @@ namespace Robust.Shared.Network
         [Dependency] private readonly ProfManager _prof = default!;
         [Dependency] private readonly HttpClientHolder _http = default!;
         [Dependency] private readonly IHWId _hwId = default!;
+        [Dependency] private readonly ITransferManager _transfer = default!;
 
         /// <summary>
         ///     Holds lookup table for NetMessage.Id -> NetMessage.Type
@@ -140,12 +142,17 @@ namespace Robust.Shared.Network
         private ISawmill _loggerPacket = default!;
         private ISawmill _authLogger = default!;
 
+        private bool _clientSerializerComplete;
+        private bool _clientTransferComplete;
+
         /// <inheritdoc />
         public int Port => _config.GetCVar(CVars.NetPort);
 
         public bool IsAuthEnabled => _config.GetCVar<bool>("auth.enabled");
 
         public IReadOnlyDictionary<Type, long> MessageBandwidthUsage => _bandwidthUsage;
+
+        internal StringTable StringTable => _strings;
 
         /// <inheritdoc />
         public bool IsServer { get; private set; }
@@ -271,6 +278,7 @@ namespace Robust.Shared.Network
             _strings.Initialize(() => { _logger.Info("Message string table loaded."); },
                 UpdateNetMessageFunctions);
             _serializer.ClientHandshakeComplete += OnSerializerOnClientHandshakeComplete;
+            _transfer.ClientHandshakeComplete += OnTransferOnClientHandshakeComplete;
 
             _initialized = true;
 
@@ -304,7 +312,21 @@ namespace Robust.Shared.Network
         private void OnSerializerOnClientHandshakeComplete()
         {
             _logger.Info("Client completed serializer handshake.");
-            OnConnected(ServerChannelImpl!);
+            _clientSerializerComplete = true;
+            ClientCheckSwitchToConnected();
+        }
+
+        private void OnTransferOnClientHandshakeComplete()
+        {
+            _logger.Info("Client completed transfer handshake.");
+            _clientTransferComplete = true;
+            ClientCheckSwitchToConnected();
+        }
+
+        private void ClientCheckSwitchToConnected()
+        {
+            if (_clientSerializerComplete && _clientTransferComplete)
+                OnConnected(ServerChannelImpl!);
         }
 
         private void SynchronizeNetTime()
@@ -427,6 +449,9 @@ namespace Robust.Shared.Network
 
             _cancelConnectTokenSource?.Cancel();
             ClientConnectState = ClientConnectionState.NotConnecting;
+
+            _clientSerializerComplete = false;
+            _clientTransferComplete = false;
         }
 
         /// <inheritdoc />
@@ -805,7 +830,9 @@ namespace Robust.Shared.Network
 
             try
             {
-                await _serializer.Handshake(channel);
+                await Task.WhenAll(
+                    _serializer.Handshake(channel),
+                    _transfer.ServerHandshake(channel));
             }
             catch (TaskCanceledException)
             {